+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/palette.module.ts
@@ -133,7 +133,9 @@ import {
   QueryWriteDateInputComponent,
   QueryWriteRaiseQueryComponent,
   QueryWriteRespondToQueryComponent,
-  ReadQueryManagementFieldComponent
+  ReadQueryManagementFieldComponent,
+  CloseQueryComponent
+
 } from './query-management';
 import { QualifyingQuestionService, QueryManagementService } from './query-management/services';
 import { ReadTextFieldComponent, WriteTextFieldComponent } from './text';
@@ -144,6 +146,7 @@ import { WaysToPayFieldComponent } from './waystopay';
 import { ReadYesNoFieldComponent, WriteYesNoFieldComponent, YesNoService } from './yes-no';
 import { QueryConfirmationComponent } from './query-management/components/query-confirmation/query-confirmation.component';
 import { QueryManagementUtils } from './query-management/utils/query-management.utils';
+import { ErrorsModule } from '../error/errors.module';
 
 const PALETTE_COMPONENTS = [
   UnsupportedFieldComponent,
@@ -273,6 +276,7 @@ const PALETTE_COMPONENTS = [
   QueryAttachmentsReadComponent,
   QueryEventCompletionComponent,
   QueryConfirmationComponent,
+  CloseQueryComponent,
 
   // Case event completion
   CaseEventCompletionComponent,
@@ -314,7 +318,8 @@ const PALETTE_COMPONENTS = [
     MatDialogModule,
     MediaViewerModule,
     LoadingModule,
-    MarkdownComponentModule
+    MarkdownComponentModule,
+    ErrorsModule
   ],
   declarations: [
     FixedListPipe,