+++ RELEASE-NOTES.md
@@ -1,5 +1,20 @@
 ## RELEASE NOTES
 
+### Version 7.2.19
+**EXUI-3301** Query Management component caches stale case ID on case switch
+
+### Version 7.2.18
+**EXUI-3147** MV CVE fix
+
+### Version 7.2.17
+**EXUI-2895** Refactor PED access token from URL
+
+### Version 7.2.16
+**EXUI-3167** Query Management - Ability to close a query thread for caseworkers
+
+### Version 7.2.15
+**EXUI-3239** QM: Concurrency error message
+
 ### Version 7.2.14
 **EXUI-3257** Users unable to upload documents - cdam fix